+++ taskflow/core/algorithm/reduce.hpp
@@ -99,7 +99,7 @@ Task FlowBuilder::reduce_guided(
               
         size_t z = s0 + 2;
         size_t p1 = 2 * W * (C + 1);
-        double p2 = 0.5 / W;
+        double p2 = 0.5 / static_cast<double>(W);
         s0 = next.load(std::memory_order_relaxed);
 
         while(s0 < N) {
@@ -124,7 +124,7 @@ Task FlowBuilder::reduce_guided(
           }
           // coarse-grained
           else {
-            size_t q = static_cast<size_t>(p2 * r);
+            size_t q = static_cast<size_t>(p2 * static_cast<double>(r));
             if(q < C) {
               q = C;
             }
@@ -526,7 +526,7 @@ Task FlowBuilder::transform_reduce_guided(
               
         size_t z = s0 + 2;
         size_t p1 = 2 * W * (C + 1);
-        double p2 = 0.5 / W;
+        double p2 = 0.5 / static_cast<double>(W);
         s0 = next.load(std::memory_order_relaxed);
 
         while(s0 < N) {
@@ -551,7 +551,7 @@ Task FlowBuilder::transform_reduce_guided(
           }
           // coarse-grained
           else {
-            size_t q = static_cast<size_t>(p2 * r);
+            size_t q = static_cast<size_t>(p2 * static_cast<double>(r));
             if(q < C) {
               q = C;
             }
